Create ConvertBinaryToOctal.swift (#5530)
This function takes a string representing a binary number as input. It first converts the binary number to a decimal using Int(initializedWith:radix:), which takes the binary string and the radix (2 for binary). Then, it converts the decimal number to an octal using String(initializedWith:radix:) with radix 8. If the input is not a valid binary number, it returns an error message.

Co-authored-by: Ritesh Kokam <61982298+RiteshK-611@users.noreply.github.com>