Create CheckValidDate.swift (#5618)
In this program, isValidDate function uses DateFormatter with the format "yyyy-MM-dd". The property isLenient is set to false to ensure the formatter does not adjust invalid dates to make them valid. For example, February 29, 2019, is an invalid date because 2019 is not a leap year, and this program will correctly output False for this input.

Co-authored-by: Ritesh Kokam <61982298+RiteshK-611@users.noreply.github.com>